Uploaded source code
There's nothing to see here - it's just code, just uploaded, not more and not less.


















Stop reading this!





























Is this really that interesting?


































I hope you're entertained...





























That's it! The end of the comment! There's no text after this line!










































Or maybe there is...

























































Just maybe...






















































































































Ok, you at least deserve something for your dedication















































I guess i can explain how the program works
















































So we enter the main() function, and declare wndHandle to be null. Since handles are just pointers it makes sense. Then we call system-defined function EnumWindows which takes a function address as an argument. From here we are doomed to enter wndEnumerator for every single window in the system that is present (at least i think so). Fun fact: each control (button, text field, image container, scroll bar, and so on) is also considered a window! When we ender wndEnumerator we get a window handle as a void* and a parameter that is kinda useless in our case. Then with the help of GetWindowText we get, surprisingly, text, and in case of program windows it's what is written in the title. Then a comparison happens, where we check if first 3 letters of the window text are SAI (case sensitive), and if so, we store the window handle, and then set the last error code to success (no errors happened) and return 0. The return value of 0 tells EnumWindows that we don't need to enumerate any more, and a non-zero value is a "continue" sign. After all that we return to main() and come to an if() statement. If the window handle is not null, which means we found one that suits our needs, we proceed to send a system message to that window. Values 0x111 and 0x102 are not random, these values trigger the same exact thing as pressing "Save" button in the menu does. I found this out by spending some time in a debugger with breakpoints set to some system functions and places where saving action takes place. After we politely told SAI to save the file, we just print a simple line of text to the console, and sleep either 5 minutes or 5 seconds, depending on the variant of code that is running. Since there's no actual exit point, and the only way to stop the program is to brutally murder it, there's no end to the cycle, and we jump back to declaring wndHandle as null.



















No, seriously though, you should stop reading this.























































Please, don't continue!





































STOP!!!


































READING!!!!!!!!!!!!











































T H I S ! ! ! ! ! !
















































GAAAARGHHHHHH!!!!!!!








































































The end.